Backend_cnn/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84
tensorflow-cpu==2.16.1
numpy==1.26.4          
pillow==10.4.0
SQLAlchemy==2.0.34
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0  
reportlab==4.2.4